ARM: omap: sched_clock: Don't let sched_clock tick during suspend
ARM: omap: sched_clock: Don't let sched_clock tick during suspend

The register used by sched_clock continues to count while in suspend.
The timer that is used to prevent wraps, which normally fires after
90% of the register wrap time, does not count in suspend.  This
causes the timer to fire after the register has wrapped, which can
cause sched_clock to wrap.

Track an offset value that is updated in suspend and resume to cause
the register read by sched_clock to appear not to count in suspend.
This will cause the sched_clock time before and after suspend to
be the same, but will prevent sched_clock wraps after 131072 seconds
(36.4 hours).

Change-Id: I37054a13232c3a4320a5a49fd99e8ab0c2d1c4fb
Signed-off-by: Colin Cross <ccross@android.com>