ARM: OMAP4: Report clock module timeouts when enabling
ARM: OMAP4: Report clock module timeouts when enabling

When a clock module is enabled, we wait for the idle-status to indicate that
the module is now enabled before continuing. Although the current code returns
an error if the idle-status does not change, we do not display any warning
message to the user. So print a message if the module idle-status does not
change as expected.

Change-Id: Ia6583174ac209f78aeecefc80cd9872d899246a2
Signed-off-by: Jon Hunter <jon-hunter@ti.com>