OMAP4: SAR: Store OMAP revision on SAR
OMAP4: SAR: Store OMAP revision on SAR

OMAP revision is a useful piece of info to have available in SAR.
At the moment it is required for POR restore on GP devices while
coming out of OFF state. POR service was added to 4430GP ROM in ES2.2
and is not available before that.

Change-Id: I7cd6a1c45d9cbe5dedd4cef39aaee88a55b117c3
Signed-off-by: Vimarsh Zutshi <vimarsh.zutshi@ti.com>